+++ src/shared/signalk/sk.worker.ts
@@ -1,5 +1,5 @@
 import * as Comlink from 'comlink';
-import type { Context, Path, SignalKClientApi, SKFrame, SubscribeEntry } from './types';
+import type { Context, Delta, Path, SignalKClientApi, SKFrame, SubscribeEntry } from './types';
 import { WorkerCore } from './worker-core';
 
 class SignalKWorker implements SignalKClientApi {
@@ -15,6 +15,10 @@ class SignalKWorker implements SignalKClientApi {
 
   async unsubscribe(paths: Path[], context?: Context): Promise<void> {
     this.#core.unsubscribe(paths, context);
+  }
+
+  async publish(delta: Delta): Promise<void> {
+    this.#core.publish(delta);
   }
 
   async disconnect(): Promise<void> {